Rename Auth to Security, that is not a just-auth configuration, it is a security configuration, and separate passwordEncoder and algorithm in a SecurityConfiguration and move SecurityConfigurerAdapter to auth package
